requirements.txt
# Core
ultralytics
fastapi
uvicorn[standard]
python-multipart
pillow
opencv-python-headless
numpy

# Face Recognition
insightface
onnxruntime

# Kafka
kafka-python